pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lab-tools"
version = "0.1.0"
description = "General tools for a computational chemistry lab"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name="Your Lab", email="YOUR_EMAIL"}]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pandas>=2.1",
  "pyarrow>=15",
  "jinja2>=3.1",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "jsonschema>=4.23",
]

[project.scripts]
labtools = "labtools.cli:app"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
